Migrating `steam-user` Version
As of 2023-09-12, Steam does not return an access token in response to successful authentication, so this won't be set when the authenticated event is fired. To address the problem, we need to bump `steam-user` version to v5 that contains breaking changes. This makes the code compatible with these changes making usage of the refreshToken.